Bug Fix
The code as written tried to build a list of existing parts, but the key for the list was not unique across different parts. The fix uses the part's persistentID, which should be unique.

What would happen is the first part would set the value of all subsequent parts in the same vessel, i.e. all reaction wheels got the same value. Not a big deal if it's a big reaction wheel. But if the first reaction wheel is a small probe core, then it nerfs the reaction wheels of the entire vessel. Also added a debug message to check the values were calculated correctly.